Move LoggedRule from ColumnarRuleApplier into ColumnarRuleExecutor
